Renamed layout files
